Authorized routes in controllers and created respective policies

+++ app/Http/Controllers/CategoryController.php
@@ -7,13 +7,15 @@ use App\Models\Category;
 use Illuminate\Http\Request;
 use App\Traits\HttpResponses;
 use App\Traits\ResourceValidator;
+use App\Traits\AuthorizationNames;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 
 class CategoryController extends Controller
 {
-    use Tables, HttpResponses, ResourceValidator;
+    use Tables, HttpResponses, ResourceValidator, AuthorizationNames;
     /**
      * Display a listing of the resource.
      */
@@ -39,6 +41,7 @@ class CategoryController extends Controller
      */
     public function store(StoreCategoryRequest $req)
     {
+        Gate::authorize($this->permNames['post-category'], Category::class);
         $req->validated();
         $category = new Category();
         $category->fill($req->validated());
@@ -71,6 +74,7 @@ class CategoryController extends Controller
     public function update(UpdateCategoryRequest $req, string $id)
     {
         $cat = Category::findOrFail($id);
+        Gate::authorize($this->permNames['put-category'], $cat);
         $cat->update($req->validated());
         $catCat = $this->categoryCategory;
         if($req->filled($catCat)) {
@@ -86,6 +90,7 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        Gate::authorize($this->permNames['delete-category'], $category);
         $category->delete();
         return $this->success($category, "Category deleted successfully");
     }
